+++ MainUI/main_menu/button.gd
@@ -12,5 +12,5 @@ func _process(_delta: float) -> void:
 
 
 func _on_button_up() -> void:
-	get_tree().change_scene_to_file("res://MainUI/combat_ui/combat_ui.tscn")
+	get_tree().change_scene_to_file("res://MainUI/combat_ui/rework/CombatRework.tscn")
 	pass # Replace with function body.